// src/server/fetch-clients/paystack/schemas.ts
export const planIntervalEnum = z.enum([
  "hourly",
  "daily",
  // "weekly",
  "monthly",
  // "quarterly",
  // "biannually",
  "annually",
]);
// ...
export const createPlanSchema = z.object({
  name: z.string().describe("Name of the plan"),
  amount: z.number().describe("Amount to be charged in subunits"),
  interval: planIntervalEnum.describe("Interval for the plan"),
  description: z.string().optional().describe("Description of the plan"),
  send_invoices: z
    .boolean()
    .optional()
    .describe(
      "Set to false if you don't want invoices to be sent to your customers",
    ),
  send_sms: z
    .boolean()
    .optional()
    .describe(
      "Set to false if you don't want text messages to be sent to your customers",
    ),
  currency: z.string().optional().describe("Currency in which amount is set"),
  invoice_limit: z
    .number()
    .optional()
    .describe("Number of invoices to raise during subscription"),
});
// ...
export const paymentPageTypeEnum = z.enum([
  "payment",
  "subscription",
  "product",
  "plan",
]);
// ...
export const createPaymentPageSchema = z.object({
  name: z.string().describe("Name of page"),
  type: paymentPageTypeEnum.optional(),
  /**The id of the plan to subscribe to */
  plan: z.string().optional(),
  split_code: z.string(),
  description: z.string().optional().describe("Description of page"),
  amount: z
    .number()
    .optional()
    .describe("Amount should be in the subunit of the supported currency"),
  slug: z
    .string()
    .optional()
    .describe("URL slug you would like to be associated with this page"),
  metadata: z
    .record(z.string(), z.unknown())
    .optional()
    .describe("Stringified JSON object of custom data"),
  redirect_url: z
    .string()
    .optional()
    .describe(
      "If you would like Paystack to redirect someplace upon successful payment, specify the URL here",
    ),
  custom_fields: z
    .array(
      z.object({
        display_name: z.string().describe("Field label"),
        variable_name: z.string().describe("Field variable name"),
        required: z
          .boolean()
          .optional()
          .describe("Set to true to make field required"),
      }),
    )
    .optional()
    .describe("If you would like to accept custom fields, specify them here"),
});